解决接口报错

+++ testcase-delivery/src/main/java/com/miller/delivery/testcase/module/deliveryAdmin/activity/ChongdanActivityListTests.java
@@ -118,7 +118,8 @@ public class ChongdanActivityListTests {
         // 9) 验证数据库中有新增的活动
         String sql = String.format("select count(0) as count from panda_test.hp_delivery_activity where `name`='自动化冲单活动%s' order by id desc ;", todayDate);
         Map<String, Object> dbResult = PandaTestDBHelpful.executeSelectOneSql(sql);
-        Integer chongdanCount = (Integer) dbResult.get("count");
+        Long chongdanCount = (Long) dbResult.get("count");
+
         assert chongdanCount != null && chongdanCount > 0 : "数据库中没有找到新增的活动";
 
         // 10) ROI-加价活动列表-传日期查询
